refactor: config/definitions -> config/options
