+++ docs/development/configuration.md
@@ -22,7 +22,7 @@ The above are listed in **_reverse order_** of preference. e.g. CLI values will 
 
 ### Default Configuration
 
-The default configuration values can be found in [lib/config/definitions.ts](../../lib/config/definitions.ts).
+The default configuration values can be found in [lib/config/options/index.ts](../../lib/config/options/index.ts).
 Options which have `"admin": true` are reserved only for bot admin configuration and cannot be configured within repository config files.
 
 ### Configuration File
@@ -75,4 +75,4 @@ If you add configuration options to your `package.json` then these will override
 
 Please see [https://docs.renovatebot.com/configuration-options/](https://docs.renovatebot.com/configuration-options/) for a list of user-facing configuration options.
 
-For further options when running your own instance of Renovate, please see the full config definitions file at `lib/config/definitions.ts`.
+For further options when running your own instance of Renovate, please see the full config options file at `lib/config/options/index.ts`.
