+++ lib/config/massage.ts
@@ -1,6 +1,6 @@
 import is from '@sindresorhus/is';
 import { clone } from '../util/clone';
-import { getOptions } from './definitions';
+import { getOptions } from './options';
 import type { PackageRule, RenovateConfig, UpdateType } from './types';
 
 const options = getOptions();
