+++ lib/config/migration.ts
@@ -4,7 +4,7 @@ import { dequal } from 'dequal';
 import { logger } from '../logger';
 import { clone } from '../util/clone';
 import { getAdminConfig } from './admin';
-import { getOptions } from './definitions';
+import { getOptions } from './options';
 import { removedPresets } from './presets/common';
 import type {
   MigratedConfig,
