+++ lib/util/template/index.spec.ts
@@ -1,5 +1,5 @@
 import { getName } from '../../../test/util';
-import { getOptions } from '../../config/definitions';
+import { getOptions } from '../../config/options';
 import * as template from '.';
 
 describe(getName(), () => {
